+++ huisbot/Models/Huis/HuisCalculationScore.cs
@@ -1,5 +1,4 @@
 ﻿using huisbot.Models.Osu;
-using huisbot.Utilities;
 using Newtonsoft.Json;
 
 namespace huisbot.Models.Huis;
