+++ huisbot/Models/Osu/OsuBeatmap.cs
@@ -1,5 +1,4 @@
-﻿using huisbot.Utilities;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
 namespace huisbot.Models.Osu;
