+++ huisbot/Modules/Huis/Feedback.cs
@@ -1,9 +1,9 @@
 ﻿using Discord;
 using Discord.Interactions;
 using Discord.WebSocket;
+using huisbot.Helpers;
 using huisbot.Models.Huis;
 using huisbot.Services;
-using huisbot.Utilities;
 
 namespace huisbot.Modules.Huis;
