+++ huisbot/Modules/Huis/Rankings.cs
@@ -1,5 +1,6 @@
 ﻿using Discord;
 using Discord.Interactions;
+using huisbot.Helpers;
 using huisbot.Models.Huis;
 using huisbot.Services;
 using huisbot.Utilities;
