+++ huisbot/Modules/Huis/Statistic.cs
@@ -1,8 +1,8 @@
 ﻿using Discord;
 using Discord.Interactions;
+using huisbot.Helpers;
 using huisbot.Models.Huis;
 using huisbot.Services;
-using huisbot.Utilities;
 using ScottPlot;
 using ScottPlot.Plottable;
 using ScottPlot.Renderable;
