+++ huisbot/Modules/Huis/TopPlays.cs
@@ -1,5 +1,6 @@
 ﻿using Discord;
 using Discord.Interactions;
+using huisbot.Helpers;
 using huisbot.Models.Huis;
 using huisbot.Models.Osu;
 using huisbot.Models.Persistence;
