+++ huisbot/Modules/Link.cs
@@ -1,5 +1,6 @@
 ﻿using Discord;
 using Discord.Interactions;
+using huisbot.Helpers;
 using huisbot.Models.Osu;
 using huisbot.Services;
 using huisbot.Utilities;
