+++ huisbot/Modules/Miscellaneous/EffMissCount.cs
@@ -1,7 +1,7 @@
 ﻿using Discord;
 using Discord.Interactions;
+using huisbot.Helpers;
 using huisbot.Models.Osu;
-using huisbot.Utilities;
 
 namespace huisbot.Modules.Miscellaneous;
