+++ huisbot/Services/InteractionHandler.cs
@@ -50,10 +50,10 @@ public class InteractionHandler(DiscordSocketClient client, ILogger<InteractionH
     }
 
     // Log the command execution.
-    string guild = command.GuildId is null ? "Unknown" : $"{Client.GetGuild(command.GuildId.Value)} ({command.GuildId})";
+    string guild = command.GuildId is null ? "Direct Message" : $"{Client.GetGuild(command.GuildId.Value)} ({command.GuildId})";
     string user = $"{command.User.Username} [{command.User.GlobalName}] ({command.User.Id})";
     string cmd = $"/{command.CommandName}{parse(command.Data.Options)}";
-    Logger.LogInformation("  Guild: {guild}\n   User: {user}\nCommand: {cmd}", guild, user, cmd);
+    Logger.LogInformation("Guild: {Guild}\n      User: {User}\n      Command: {Command}", guild, user, cmd);
 
     return Task.CompletedTask;
   }
